+++ packages/export/src/inject-state.js
@@ -4,11 +4,12 @@ const log = require('@home-gallery/logger')('export.webapp.injectState');
 
 const rewriteFile = require('./rewrite-file');
 
-const injectState = (database, outputDirectory, basePath, cb) => {
+const injectState = (database, outputDirectory, basePath, disableEdit, cb) => {
   const t0 = Date.now();
   const indexFilename = path.join(outputDirectory, basePath, 'index.html')
   rewriteFile(indexFilename, data => {
     const state = {
+      disabled: disableEdit ? ['edit', 'serverEvent'] : [],
       entries: database.data.slice(0, 50)
     };
     return data.replace('window.__homeGallery={}', `window.__homeGallery=${JSON.stringify(state)}`);