+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- export: Add --edit option to enable edit menu
 - fetch: Add --delete option to remove non existing local files
 - server: Add base path option
 
 ### Changed
 
+- export: Disable edit menu by default. Use --edit to enable it
 - webapp: Upgrade react router
 - Replace lerna by pnpm to build the project
 - Upgrade to node 18